Fix nightly clippy warnings (#179)

+++ poc_mobile_verifier/src/verifier.rs
@@ -174,7 +174,7 @@ async fn last_verified_end_time(exec: impl PgExecutor<'_>) -> Result<DateTime<Ut
 }
 
 async fn save_last_verified_end_time(exec: impl PgExecutor<'_>, value: &DateTime<Utc>) -> Result {
-    meta::store(exec, "last_verified_end_time", value.timestamp() as i64)
+    meta::store(exec, "last_verified_end_time", value.timestamp())
         .await
         .map_err(Error::from)
 }
@@ -184,7 +184,7 @@ async fn last_rewarded_end_time(exec: impl PgExecutor<'_>) -> Result<DateTime<Ut
 }
 
 async fn save_last_rewarded_end_time(exec: impl PgExecutor<'_>, value: &DateTime<Utc>) -> Result {
-    meta::store(exec, "last_rewarded_end_time", value.timestamp() as i64)
+    meta::store(exec, "last_rewarded_end_time", value.timestamp())
         .await
         .map_err(Error::from)
 }
@@ -194,7 +194,7 @@ async fn next_rewarded_end_time(exec: impl PgExecutor<'_>) -> Result<DateTime<Ut
 }
 
 async fn save_next_rewarded_end_time(exec: impl PgExecutor<'_>, value: &DateTime<Utc>) -> Result {
-    meta::store(exec, "next_rewarded_end_time", value.timestamp() as i64)
+    meta::store(exec, "next_rewarded_end_time", value.timestamp())
         .await
         .map_err(Error::from)
 }
